Move method `_elim_aliases` to `kore.manip`
